// libs/flowise-client/src/t-config.d.ts
export type TFlowiseClientConfig = {
    apiUrl: string;
    apiKey: string;
    requestTimeoutMs?: number;
    throttleMs?: number;
    maxRetries?: number;
};
export declare const DEFAULT_REQUEST_TIMEOUT_MS = 30000;
export declare const DEFAULT_THROTTLE_MS = 50;
export declare const DEFAULT_MAX_RETRIES = 3;
//# sourceMappingURL=t-config.d.ts.map